VSL-131: Added reusable component
