+++ packages/client/src/components/SafeSettings/components/Assets/index.js
@@ -1,22 +1,21 @@
-import React, { useContext, useEffect, useState } from "react";
+import React, { useContext, useEffect } from "react";
 import { useHistory } from "react-router-dom";
 import { useModalContext } from "contexts";
-import { formatAddress, getFlowscanUrlForContract } from "utils";
-import { useClipboard } from "../../../../hooks";
+import { formatAddress } from "utils";
 
 import { Web3Context } from "contexts/Web3";
-import { Plus, Copy } from "components/Svg";
+import { Plus } from "components/Svg";
 import AddVault from "./AddVault";
 import AddCollection from "./AddCollection";
+import AssetTableView from "./AssetTableView";
 
 const Assets = ({ treasury, proposeAddVault, proposeAddCollection }) => {
   const { address } = treasury;
   const { openModal, closeModal } = useModalContext();
   const history = useHistory();
-  const clipboard = useClipboard();
 
   const { getTreasuryCollections, getTreasuryVaults, vaults, NFTs } = useContext(Web3Context);
-  
+
   useEffect(() => {
     if (!address) {
       return;
@@ -72,33 +71,7 @@ const Assets = ({ treasury, proposeAddVault, proposeAddCollection }) => {
         </button>
       </div>
       {vaults && vaults[address] &&
-        <div className="border-light rounded-sm py-3 table-border">
-          {Object.values(vaults[address]).map(vault => {
-            const address = vault.split(".")[1];
-            const formattedAddress = formatAddress(address);
-            const name = vault.split(".")[2];
-            return (
-              <div className="mx-0 p-3 columns">
-                <div className="column">{name}</div>
-                <div className="column has-text-weight-bold">
-                  {formattedAddress}
-                  <span className="pointer" onClick={() => clipboard.copy(formattedAddress)}>
-                    <Copy className="mt-1 ml-2 pointer" />
-                  </span></div>
-                <a
-                  className="column is-flex is-justify-content-end has-text-purple button border-none"
-                  href={getFlowscanUrlForContract(address, name)}
-                  target="_blank"
-                  rel="noreferrer">
-                  View on Flowscan
-                </a>
-                <button className="column is-flex is-justify-content-end has-text-purple button border-none">
-                  Remove
-                </button>
-              </div>
-            )
-          })}
-        </div>
+        <AssetTableView assets={Object.values(vaults[address])} />
       }
 
       <div className="p-0 mt-5 columns">
@@ -112,34 +85,7 @@ const Assets = ({ treasury, proposeAddVault, proposeAddCollection }) => {
         </button>
       </div>
       {NFTs && NFTs[address] &&
-        <div className="border-light rounded-sm py-3 table-border">
-          {Object.keys(NFTs[address]).map(collection => {
-            const address = collection.split(".")[1];
-            const formattedAddress = formatAddress(address);
-            const name = collection.split(".")[2];
-            return (
-              <div className="mx-0 p-3 columns">
-                <div className="column">{name}</div>
-                <div className="column has-text-weight-bold">
-                  {formattedAddress}
-                  <span className="pointer" onClick={() => clipboard.copy(formattedAddress)}>
-                    <Copy className="mt-1 ml-2 pointer" />
-                  </span>
-                </div>
-                <a
-                  className="column is-flex is-justify-content-end has-text-purple button border-none"
-                  href={getFlowscanUrlForContract(address, name)}
-                  target="_blank"
-                  rel="noreferrer">
-                  View on Flowscan
-                </a>
-                <button className="column is-flex is-justify-content-end has-text-purple button border-none">
-                  Remove
-                </button>
-              </div>
-            );
-          })}
-        </div>
+        <AssetTableView assets={Object.keys(NFTs[address])} />
       }
     </div>
   );
