+++ packages/client/src/components/Svg.js
@@ -204,5 +204,5 @@ export const Copy = ({ width = "12", height = "12", className = "" }) =>
     fill="none"
     viewBox="0 0 12 12"
   >
-    <path fill-rule="evenodd" clipRule="evenodd" d="M3.375 1.5H3.35032H3.35029C3.01787 1.49998 2.73145 1.49997 2.50208 1.5308C2.25626 1.56385 2.02063 1.63838 1.82951 1.82951C1.63838 2.02063 1.56385 2.25626 1.5308 2.50208C1.49997 2.73145 1.49998 3.01787 1.5 3.35029V3.35032V3.375V6.02469V6.02471C1.49998 6.35713 1.49997 6.64356 1.5308 6.87292C1.56385 7.11874 1.63838 7.35437 1.82951 7.5455C2.02063 7.73662 2.25626 7.81115 2.50208 7.8442C2.73145 7.87504 3.01789 7.87502 3.35032 7.875H3.35032L3.5625 7.875V7.125H3.375C3.01085 7.125 2.77546 7.12421 2.60202 7.10089C2.43936 7.07902 2.38838 7.04371 2.35984 7.01517C2.33129 6.98662 2.29598 6.93564 2.27412 6.77299C2.2508 6.59955 2.25 6.36416 2.25 6V3.375C2.25 3.01085 2.2508 2.77546 2.27412 2.60202C2.29598 2.43936 2.33129 2.38838 2.35984 2.35984C2.38838 2.33129 2.43936 2.29598 2.60202 2.27412C2.77546 2.2508 3.01085 2.25 3.375 2.25H6C6.36416 2.25 6.59955 2.2508 6.77299 2.27412C6.93564 2.29598 6.98662 2.33129 7.01517 2.35984C7.04371 2.38838 7.07902 2.43936 7.10089 2.60202C7.12421 2.77546 7.125 3.01085 7.125 3.375V3.5625H7.875L7.875 3.35032V3.35032C7.87502 3.01789 7.87504 2.73145 7.8442 2.50208C7.81115 2.25626 7.73662 2.02063 7.5455 1.82951C7.35437 1.63838 7.11874 1.56385 6.87292 1.5308C6.64356 1.49997 6.35713 1.49998 6.02471 1.5H6.02469H3.375ZM6.75 4.875L6.72532 4.875C6.39288 4.87498 6.10645 4.87497 5.87708 4.9058C5.63126 4.93885 5.39563 5.01338 5.20451 5.20451C5.01338 5.39563 4.93885 5.63126 4.9058 5.87708C4.87497 6.10645 4.87498 6.39288 4.875 6.72532L4.875 6.75L4.875 9.39969C4.87498 9.73212 4.87497 10.0186 4.9058 10.2479C4.93885 10.4937 5.01338 10.7294 5.20451 10.9205C5.39563 11.1116 5.63126 11.1861 5.87708 11.2192C6.10645 11.25 6.39288 11.25 6.7253 11.25H6.72532H9.39968H9.3997C9.73213 11.25 10.0186 11.25 10.2479 11.2192C10.4937 11.1861 10.7294 11.1116 10.9205 10.9205C11.1116 10.7294 11.1861 10.4937 11.2192 10.2479C11.25 10.0186 11.25 9.73213 11.25 9.3997V9.39968V6.72532V6.7253C11.25 6.39288 11.25 6.10645 11.2192 5.87708C11.1861 5.63126 11.1116 5.39563 10.9205 5.20451C10.7294 5.01338 10.4937 4.93885 10.2479 4.9058C10.0186 4.87497 9.73212 4.87498 9.39969 4.875L6.75 4.875ZM5.73484 5.73484C5.76338 5.70629 5.81436 5.67098 5.97702 5.64912C6.15046 5.6258 6.38585 5.625 6.75 5.625H9.375C9.73916 5.625 9.97455 5.6258 10.148 5.64912C10.3106 5.67098 10.3616 5.70629 10.3902 5.73484C10.4187 5.76338 10.454 5.81436 10.4759 5.97702C10.4992 6.15046 10.5 6.38585 10.5 6.75V9.375C10.5 9.73916 10.4992 9.97455 10.4759 10.148C10.454 10.3106 10.4187 10.3616 10.3902 10.3902C10.3616 10.4187 10.3106 10.454 10.148 10.4759C9.97455 10.4992 9.73916 10.5 9.375 10.5H6.75C6.38585 10.5 6.15046 10.4992 5.97702 10.4759C5.81436 10.454 5.76338 10.4187 5.73484 10.3902C5.70629 10.3616 5.67098 10.3106 5.64912 10.148C5.6258 9.97455 5.625 9.73916 5.625 9.375V6.75C5.625 6.38585 5.6258 6.15046 5.64912 5.97702C5.67098 5.81436 5.70629 5.76338 5.73484 5.73484Z" fill="black" />
+    <path fillRule="evenodd" clipRule="evenodd" d="M3.375 1.5H3.35032H3.35029C3.01787 1.49998 2.73145 1.49997 2.50208 1.5308C2.25626 1.56385 2.02063 1.63838 1.82951 1.82951C1.63838 2.02063 1.56385 2.25626 1.5308 2.50208C1.49997 2.73145 1.49998 3.01787 1.5 3.35029V3.35032V3.375V6.02469V6.02471C1.49998 6.35713 1.49997 6.64356 1.5308 6.87292C1.56385 7.11874 1.63838 7.35437 1.82951 7.5455C2.02063 7.73662 2.25626 7.81115 2.50208 7.8442C2.73145 7.87504 3.01789 7.87502 3.35032 7.875H3.35032L3.5625 7.875V7.125H3.375C3.01085 7.125 2.77546 7.12421 2.60202 7.10089C2.43936 7.07902 2.38838 7.04371 2.35984 7.01517C2.33129 6.98662 2.29598 6.93564 2.27412 6.77299C2.2508 6.59955 2.25 6.36416 2.25 6V3.375C2.25 3.01085 2.2508 2.77546 2.27412 2.60202C2.29598 2.43936 2.33129 2.38838 2.35984 2.35984C2.38838 2.33129 2.43936 2.29598 2.60202 2.27412C2.77546 2.2508 3.01085 2.25 3.375 2.25H6C6.36416 2.25 6.59955 2.2508 6.77299 2.27412C6.93564 2.29598 6.98662 2.33129 7.01517 2.35984C7.04371 2.38838 7.07902 2.43936 7.10089 2.60202C7.12421 2.77546 7.125 3.01085 7.125 3.375V3.5625H7.875L7.875 3.35032V3.35032C7.87502 3.01789 7.87504 2.73145 7.8442 2.50208C7.81115 2.25626 7.73662 2.02063 7.5455 1.82951C7.35437 1.63838 7.11874 1.56385 6.87292 1.5308C6.64356 1.49997 6.35713 1.49998 6.02471 1.5H6.02469H3.375ZM6.75 4.875L6.72532 4.875C6.39288 4.87498 6.10645 4.87497 5.87708 4.9058C5.63126 4.93885 5.39563 5.01338 5.20451 5.20451C5.01338 5.39563 4.93885 5.63126 4.9058 5.87708C4.87497 6.10645 4.87498 6.39288 4.875 6.72532L4.875 6.75L4.875 9.39969C4.87498 9.73212 4.87497 10.0186 4.9058 10.2479C4.93885 10.4937 5.01338 10.7294 5.20451 10.9205C5.39563 11.1116 5.63126 11.1861 5.87708 11.2192C6.10645 11.25 6.39288 11.25 6.7253 11.25H6.72532H9.39968H9.3997C9.73213 11.25 10.0186 11.25 10.2479 11.2192C10.4937 11.1861 10.7294 11.1116 10.9205 10.9205C11.1116 10.7294 11.1861 10.4937 11.2192 10.2479C11.25 10.0186 11.25 9.73213 11.25 9.3997V9.39968V6.72532V6.7253C11.25 6.39288 11.25 6.10645 11.2192 5.87708C11.1861 5.63126 11.1116 5.39563 10.9205 5.20451C10.7294 5.01338 10.4937 4.93885 10.2479 4.9058C10.0186 4.87497 9.73212 4.87498 9.39969 4.875L6.75 4.875ZM5.73484 5.73484C5.76338 5.70629 5.81436 5.67098 5.97702 5.64912C6.15046 5.6258 6.38585 5.625 6.75 5.625H9.375C9.73916 5.625 9.97455 5.6258 10.148 5.64912C10.3106 5.67098 10.3616 5.70629 10.3902 5.73484C10.4187 5.76338 10.454 5.81436 10.4759 5.97702C10.4992 6.15046 10.5 6.38585 10.5 6.75V9.375C10.5 9.73916 10.4992 9.97455 10.4759 10.148C10.454 10.3106 10.4187 10.3616 10.3902 10.3902C10.3616 10.4187 10.3106 10.454 10.148 10.4759C9.97455 10.4992 9.73916 10.5 9.375 10.5H6.75C6.38585 10.5 6.15046 10.4992 5.97702 10.4759C5.81436 10.454 5.76338 10.4187 5.73484 10.3902C5.70629 10.3616 5.67098 10.3106 5.64912 10.148C5.6258 9.97455 5.625 9.73916 5.625 9.375V6.75C5.625 6.38585 5.6258 6.15046 5.64912 5.97702C5.67098 5.81436 5.70629 5.76338 5.73484 5.73484Z" fill="black" />
   </svg>
